+++ apps/web/next.config.js
@@ -1,3 +1,5 @@
+const packageJson = require('./package.json');
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     reactStrictMode: true,
@@ -8,6 +10,12 @@ const nextConfig = {
     },
     images: {
         domains: ['iwxokvfmyyfxwusixqcc.supabase.co'],
+    },
+    devIndicators: {
+        appIsrStatus: false,
+    },
+    env: {
+        NEXT_PUBLIC_APP_VERSION: packageJson.version,
     },
 }